hop-node: Seperate out available liquidity methods to own class
